+++ faq-bundle/src/Resources/contao/config/config.php
@@ -49,7 +49,6 @@ if (TL_MODE == 'BE')
  */
 $GLOBALS['TL_HOOKS']['getSearchablePages'][] = array('ModuleFaq', 'getSearchablePages');
 $GLOBALS['TL_HOOKS']['replaceInsertTags'][] = array('contao_faq.listener.insert_tags', 'onReplaceInsertTags');
-$GLOBALS['TL_HOOKS']['addFileMetaInformationToRequest'][] = array('contao_faq.listener.file_meta_information', 'onAddFileMetaInformationToRequest');
 
 
 /**